pyproject.toml
[project]
name = "career-chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "python-dotenv",
    "gradio",
    "pypdf",
    "boto3",
    "strands-agents",
]
